wip(#1101): extract lower/module_decl
